Improvements acquisitions with MPI.
Proxy the detector MPI communicator to the instrument instance.
When indexing the detectors (or the samplings) of an acquisition, change the communicator of the detectors (or the samplings) to MPI.COMM_SELF.
Fix clearing the cached operator when indexing the acquisition.
Add MPI-specific tests and execute them with mpirun in the CI.